SWARM-1668 rename artifacts and package
