--- implementation/src/main/java/org/wildfly/swarm/microprofile/fault/tolerance/hystrix/State.java
@@ -1,16 +1,0 @@
-package org.wildfly.swarm.microprofile.fault.tolerance.hystrix;
-
-/**
- *
- */
-public interface State {
-    boolean allowsExecution(int execCount);
-
-    SynchronousCircuitBreaker.Status getState();
-
-    default void recordFailure() {
-    }
-
-    default void recordSuccess() {
-    }
-}
