+++ implementation/src/main/java/org/wildfly/swarm/microprofile/faulttolerance/State.java
@@ -1,0 +1,16 @@
+package org.wildfly.swarm.microprofile.faulttolerance;
+
+/**
+ *
+ */
+public interface State {
+    boolean allowsExecution(int execCount);
+
+    SynchronousCircuitBreaker.Status getState();
+
+    default void recordFailure() {
+    }
+
+    default void recordSuccess() {
+    }
+}
