+++ tck-runner/src/test/resources/META-INF/services/org.jboss.arquillian.core.spi.LoadableExtension
@@ -1,1 +1,1 @@
-org.wildfly.swarm.microprofile.fault.tolerance.hystrix.TCKIntegrationExtension
+org.wildfly.swarm.microprofile.faulttolerance.TCKIntegrationExtension
